requirements.txt
py2neo
fuzzywuzzy
python-Levenshtein
numpy
scipy
pandas
#seaborn
#matplotlib
gensim
flask
werkzeug
tables
elasticsearch
Flask-Cors
flask_sqlalchemy
psycopg2-binary
pyyaml
webargs
flask-marshmallow
BaseHash
